Java #7: nextFloat returns a float now.

+++ vol1/java-examples/src/main/java/com/heatonresearch/aifh/randomize/BasicGenerateRandom.java
@@ -77,7 +77,7 @@ public class BasicGenerateRandom extends AbstractGenerateRandom {
      * {@inheritDoc}
      */
     @Override
-    public double nextFloat() {
+    public float nextFloat() {
         return this.rand.nextFloat();
     }
