+++ vol1/java-examples/src/main/java/com/heatonresearch/aifh/randomize/MersenneTwisterGenerateRandom.java
@@ -161,7 +161,7 @@ public class MersenneTwisterGenerateRandom extends AbstractBoxMuller {
      * {@inheritDoc}
      */
     @Override
-    public double nextFloat() {
+    public float nextFloat() {
         return (float) nextDouble();
     }
